requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# API clients
requests>=2.31.0
httpx>=0.27.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Logging and monitoring
structlog>=23.0.0
rich>=13.0.0

# Email and communication
sendgrid>=6.10.0

# Google Sheets integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Data enrichment APIs
# clearbit>=0.1.7

# Async support
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development tools
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
